tests/unit/rag_service/test_rag_retrieval.py
"""
Unit Tests for RAG (Retrieval-Augmented Generation) Service

BUSINESS REQUIREMENT:
Tests RAG system for contextual learning content retrieval and AI-enhanced
educational content generation based on student progress and learning patterns.

TECHNICAL IMPLEMENTATION:
Tests vector embeddings, semantic search, context retrieval, and LLM integration.
"""

import pytest
from datetime import datetime
from uuid import uuid4
import numpy as np

import sys
from pathlib import Path
sys.path.insert(0, str(Path(__file__).parent.parent.parent.parent / 'services' / 'rag-service'))


class TestVectorEmbeddings:
    """Test vector embedding generation and storage"""

    def test_generate_text_embedding(self):
        """Test generating embedding vector from text"""
        text = "Python is a high-level programming language"

        # Simulate embedding generation (would use actual model in production)
        embedding = np.random.rand(384)  # Typical embedding dimension

        assert len(embedding) == 384
        assert isinstance(embedding, np.ndarray)

    def test_embedding_similarity(self):
        """Test cosine similarity between embeddings"""
        embedding1 = np.random.rand(384)
        embedding2 = np.random.rand(384)

        # Calculate cosine similarity
        similarity = np.dot(embedding1, embedding2) / (
            np.linalg.norm(embedding1) * np.linalg.norm(embedding2)
        )

        assert -1 <= similarity <= 1

    def test_normalize_embedding(self):
        """Test embedding normalization"""
        embedding = np.random.rand(384)
        normalized = embedding / np.linalg.norm(embedding)

        # Normalized vector should have unit length
        norm = np.linalg.norm(normalized)
        assert abs(norm - 1.0) < 1e-6


class TestSemanticSearch:
    """Test semantic search functionality"""

    def test_retrieve_similar_content(self):
        """Test retrieving semantically similar content"""
        query_text = "How to write a Python function"

        # Simulate document database
        documents = [
            {"id": "1", "text": "Python functions are defined using def keyword"},
            {"id": "2", "text": "Variables in Python are dynamically typed"},
            {"id": "3", "text": "Functions in Python can return multiple values"}
        ]

        # In real implementation, would compare embeddings
        # For test, simple keyword matching
        relevant_docs = [
            doc for doc in documents
            if "function" in doc["text"].lower() or "python" in doc["text"].lower()
        ]

        assert len(relevant_docs) >= 2
        assert any("function" in doc["text"].lower() for doc in relevant_docs)

    def test_rank_search_results(self):
        """Test ranking search results by relevance"""
        results = [
            {"id": "1", "score": 0.95},
            {"id": "2", "score": 0.87},
            {"id": "3", "score": 0.92}
        ]

        # Sort by score descending
        ranked = sorted(results, key=lambda x: x["score"], reverse=True)

        assert ranked[0]["score"] >= ranked[1]["score"]
        assert ranked[1]["score"] >= ranked[2]["score"]
        assert ranked[0]["id"] == "1"


class TestContextRetrieval:
    """Test context retrieval for LLM"""

    def test_retrieve_relevant_context(self):
        """Test retrieving relevant context for a query"""
        student_query = "Explain Python decorators"
        max_context_items = 3

        # Simulate knowledge base
        knowledge_base = [
            {"topic": "decorators", "content": "Decorators modify function behavior"},
            {"topic": "functions", "content": "Functions are first-class objects"},
            {"topic": "classes", "content": "Classes define object blueprints"}
        ]

        # Retrieve relevant items
        context = [
            item for item in knowledge_base
            if "decorator" in item["topic"] or "function" in item["topic"]
        ][:max_context_items]

        assert len(context) <= max_context_items
        assert len(context) >= 1

    def test_context_window_limit(self):
        """Test respecting context window limits"""
        max_tokens = 2000

        context_items = [
            "Item 1 " * 500,  # ~500 tokens
            "Item 2 " * 500,
            "Item 3 " * 500,
            "Item 4 " * 500
        ]

        # Simulate token counting
        total_tokens = 0
        selected_context = []

        for item in context_items:
            item_tokens = len(item.split())
            if total_tokens + item_tokens <= max_tokens:
                selected_context.append(item)
                total_tokens += item_tokens
            else:
                break

        assert total_tokens <= max_tokens
        assert len(selected_context) < len(context_items)


class TestProgressiveRetrieval:
    """Test progressive learning content retrieval"""

    def test_retrieve_based_on_student_level(self):
        """Test retrieving content appropriate for student level"""
        student_level = "beginner"

        content_library = [
            {"level": "beginner", "topic": "Python Basics"},
            {"level": "intermediate", "topic": "OOP in Python"},
            {"level": "advanced", "topic": "Metaclasses"}
        ]

        appropriate_content = [
            item for item in content_library
            if item["level"] == student_level
        ]

        assert len(appropriate_content) >= 1
        assert all(item["level"] == "beginner" for item in appropriate_content)

    def test_retrieve_next_learning_step(self):
        """Test retrieving next appropriate learning content"""
        completed_topics = ["Python Basics", "Variables", "Functions"]

        learning_path = [
            {"topic": "Python Basics", "order": 1},
            {"topic": "Functions", "order": 2},
            {"topic": "Classes", "order": 3},
            {"topic": "OOP", "order": 4}
        ]

        # Find next uncompleted topic
        next_topic = None
        for item in sorted(learning_path, key=lambda x: x["order"]):
            if item["topic"] not in completed_topics:
                next_topic = item
                break

        assert next_topic is not None
        assert next_topic["topic"] == "Classes"


class TestLLMIntegration:
    """Test LLM integration for content generation"""

    def test_generate_personalized_explanation(self):
        """Test generating personalized explanations"""
        student_context = {
            "level": "beginner",
            "completed_topics": ["variables", "loops"],
            "learning_style": "visual"
        }

        query = "Explain Python functions"

        # Simulate prompt construction
        prompt = f"""
        Student Level: {student_context['level']}
        Learning Style: {student_context['learning_style']}
        Query: {query}

        Provide an explanation suitable for this student.
        """

        assert "beginner" in prompt
        assert "visual" in prompt
        assert query in prompt

    def test_generate_practice_problems(self):
        """Test generating practice problems"""
        topic = "Python lists"
        difficulty = "easy"
        count = 3

        # Simulate problem generation
        problems = []
        for i in range(count):
            problems.append({
                "id": i + 1,
                "topic": topic,
                "difficulty": difficulty,
                "question": f"Problem {i + 1} about {topic}"
            })

        assert len(problems) == count
        assert all(p["topic"] == topic for p in problems)


class TestCaching:
    """Test caching for performance optimization"""

    def test_cache_embeddings(self):
        """Test caching generated embeddings"""
        cache = {}

        text = "Python programming language"
        cache_key = hash(text)

        # Generate and cache embedding
        if cache_key not in cache:
            embedding = np.random.rand(384)
            cache[cache_key] = embedding

        # Retrieve from cache
        cached_embedding = cache[cache_key]

        assert cache_key in cache
        assert len(cached_embedding) == 384

    def test_cache_search_results(self):
        """Test caching search results"""
        search_cache = {}

        query = "Python functions"
        cache_key = hash(query)

        # Cache search results
        search_results = [
            {"id": "1", "text": "Functions in Python"},
            {"id": "2", "text": "Python function syntax"}
        ]

        search_cache[cache_key] = {
            "results": search_results,
            "timestamp": datetime.utcnow()
        }

        # Retrieve from cache
        cached_results = search_cache[cache_key]["results"]

        assert len(cached_results) == 2
        assert cache_key in search_cache


class TestRelevanceFeedback:
    """Test learning from user feedback"""

    def test_record_helpful_content(self):
        """Test recording when content is helpful"""
        feedback_log = []

        content_id = "content-123"
        student_id = "student-456"
        is_helpful = True

        feedback_log.append({
            "content_id": content_id,
            "student_id": student_id,
            "helpful": is_helpful,
            "timestamp": datetime.utcnow()
        })

        assert len(feedback_log) == 1
        assert feedback_log[0]["helpful"] is True

    def test_adjust_rankings_based_on_feedback(self):
        """Test adjusting content rankings based on feedback"""
        content_scores = {
            "content-1": {"base_score": 0.8, "helpful_count": 5, "not_helpful_count": 1},
            "content-2": {"base_score": 0.7, "helpful_count": 2, "not_helpful_count": 3}
        }

        # Calculate adjusted scores
        for content_id, scores in content_scores.items():
            helpful_ratio = scores["helpful_count"] / (scores["helpful_count"] + scores["not_helpful_count"])
            adjusted_score = scores["base_score"] * (0.5 + 0.5 * helpful_ratio)
            scores["adjusted_score"] = adjusted_score

        # Content 1 should have higher adjusted score
        assert content_scores["content-1"]["adjusted_score"] > content_scores["content-2"]["adjusted_score"]


if __name__ == "__main__":
    pytest.main([__file__, "-v"])
