tests/test_requirements.txt
# Testing dependencies for Course Creator Platform

# Core testing framework
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-html>=3.2.0
pytest-xdist>=3.3.0
pytest-mock>=3.11.0

# HTTP testing
requests>=2.31.0
httpx>=0.24.0
responses>=0.23.0

# Database testing (REAL database connections, no mocking!)
pytest-postgresql>=5.0.0
asyncpg>=0.28.0
psycopg2-binary>=2.9.7  # For real PostgreSQL connections
docker>=6.1.0           # For Docker integration testing
redis>=4.6.0            # For real Redis connections

# Security testing
bandit>=1.7.5
safety>=2.3.0

# Performance testing
pytest-benchmark>=4.0.0
locust>=2.15.0

# Code quality
black>=23.7.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.5.0

# API testing
fastapi[all]>=0.103.0
starlette>=0.27.0

# Authentication testing
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# E2E testing with Selenium and Chrome driver
selenium>=4.15.0
webdriver-manager>=4.0.1  # Automatic ChromeDriver management
# playwright>=1.37.0  # Keeping for backwards compatibility

# Utilities
faker>=19.3.0
factory-boy>=3.3.0
freezegun>=1.2.2
coverage>=7.2.0

# Environment management
python-dotenv>=1.0.0

# Time and date testing
arrow>=1.2.3

# Mock services
responses>=0.23.0
httpretty>=1.1.4

# Database migration testing
alembic>=1.11.0

# Configuration testing (to catch config bugs!)
pydantic>=2.0.0
omegaconf>=2.3.0
PyYAML>=6.0.1           # For YAML config validation

# Logging and monitoring
structlog>=23.1.0

# File handling testing
aiofiles>=23.1.0

# UUID and other utilities
uuid>=1.30